+++ src/main/java/com/vaded/ae2patternrestocker/mixin/ScreenAccessor.java
@@ -5,10 +5,16 @@ import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.client.gui.narration.NarratableEntry;
 import net.minecraft.client.gui.screens.Screen;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
 import org.spongepowered.asm.mixin.gen.Invoker;
+
+import java.util.List;
 
 @Mixin(Screen.class)
 public interface ScreenAccessor {
     @Invoker("addRenderableWidget")
     <T extends GuiEventListener & Renderable & NarratableEntry> T ae2PatternRestocker$addRenderableWidget(T widget);
+
+    @Accessor("renderables")
+    List<Renderable> ae2PatternRestocker$getRenderables();
 }